+++ src/components/ProductCard.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import shoppingCartIconBlack from "../asset/shopping-cart-icon-black.svg";
+import AddToCartButton from "./AddToCartButton";
 
 const Container = styled.div`
   display: flex;
@@ -57,28 +57,6 @@ const ProductPrice = styled.p`
   font-size: 1.25rem;
   color: ${({ theme }) => theme.textDefault};
 `;
-
-const IconButton = styled.button`
-  padding: 10px 0 10px 10px;
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-
-  img {
-    width: 30px;
-    :hover {
-      transform: scale(1.05);
-    }
-  }
-`;
-
-function AddToCartButton() {
-  return (
-    <IconButton>
-      <img src={shoppingCartIconBlack} alt="장바구니 담기" />
-    </IconButton>
-  );
-}
 
 function ProductCard({ productInfo: { thumbnail, name, price } }) {
   return (
